requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
django-debug-toolbar==4.2.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangorestframework-stubs==3.14.4
drf-spectacular==0.26.5
Faker==20.1.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
multidict==6.0.4
mypy==1.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pip-autoremove==0.10.0
platformdirs==4.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.3
